# vulkan: Enable VK_KHR_cooperative_matrix extension for Intel Xe2 GPUs