Enabling VK_KHR_cooperative_matrix on Intel Xe2 GPUs (currently Lunar Lake and Battlemage) have significant performance improvement, while we also see performance regressions with older GPUs like Arc A770.
This PR will enable VK_KHR_cooperative_matrix only for Xe2 GPUs until performance regression is resolved for older GPUs.

Reference: https://github.com/ggml-org/llama.cpp/discussions/13530

# llama-bench results
All results on Windows 11 24H2 with graphics driver 32.0.101.6795.
- U7-268V
  - About 2.8x performance improvement in pp512.
- Arc B580
  - About 3.3x performance improvement in pp512.
- Arc A770
  - VK_KHR_cooperative_matrix is not enabled so no change in performance (as expected).
## Lunar Lake Core Ultra 7 268V

### Before
```
λ llama-bench.exe -m ..\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) Graphics (Intel Corporation) | uma: 1 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: none
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |        148.60 ± 3.52 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |         36.87 ± 0.61 |

build: 7e00e60e (5583)
```
### After
```
λ llama-bench.exe -m ..\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) Graphics (Intel Corporation) | uma: 1 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: KHR_coopmat
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |       416.53 ± 25.15 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |         37.86 ± 0.19 |

build: a0dd7795 (5589)
```
## Battlemage Arc B580

### Before
```
λ llama-bench.exe -m ..\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) B580 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: none
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |       490.89 ± 12.28 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |        129.95 ± 0.27 |

build: 7e00e60e (5583)
```
### After
```
λ llama-bench.exe -m ..\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) B580 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: KHR_coopmat
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |       1631.44 ± 4.42 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |        126.69 ± 0.50 |

build: a0dd7795 (5589)
```
## Alchemist Arc A770
### Before
```
λ llama-bench.exe -m C:\Users\cpie-ace\Documents\Axell\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) A770 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: none
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |        974.92 ± 4.06 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |         96.42 ± 0.40 |

build: 7e00e60e (5583)
```
### After
```
λ llama-bench.exe -m C:\Users\cpie-ace\Documents\Axell\gemma-2-2b-it-Q4_K_M.gguf
ggml_vulkan: Found 1 Vulkan devices:
ggml_vulkan: 0 = Intel(R) Arc(TM) A770 Graphics (Intel Corporation) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 32768 | int dot: 0 | matrix cores: none
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           pp512 |        977.81 ± 1.79 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | Vulkan     |  99 |           tg128 |         96.83 ± 0.30 |

build: a0dd7795 (5589)
```