@rillomas 2025-06-04
Thanks for your review. I'll benchmark the current version on Ubuntu and update the PR