@0cc4m 2025-06-05
Thank you for testing with Ubuntu. There is no reason not to merge, but it's seriously disappointing how much worse the Linux driver currently is. I hope Intel closes that gap in the near future.