@codecnotsupported 2025-06-05
> Thank you for testing with Ubuntu. There is no reason not to merge, but it's seriously disappointing how much worse the Linux driver currently is. I hope Intel closes that gap in the near future.

I do believe there is a issue in the mesa bug tracking for that exact problem.
That said, SYCL backend has better performance.

```
$ ./build/bin/llama-bench -m ./models/gemma-2-2b-it-q4_k_m.gguf 
warning: asserts enabled, performance may be affected
register_backend: registered backend SYCL (1 devices)
register_device: registered device SYCL0 (Intel(R) Arc(TM) B580 Graphics)
register_backend: registered backend CPU (1 devices)
register_device: registered device CPU (11th Gen Intel(R) Core(TM) i5-11400F @ 2.60GHz)
load_backend: failed to find ggml_backend_init in ./llama.cpp/build/bin/libggml-sycl.so
load_backend: failed to find ggml_backend_init in ./llama.cpp/build/bin/libggml-cpu.so
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | SYCL       |  99 |           pp512 |      2186.80 ± 16.75 |
| gemma2 2B Q4_K - Medium        |   1.59 GiB |     2.61 B | SYCL       |  99 |           tg128 |         54.23 ± 0.31 |
```