@0cc4m 2025-06-05
Yeah, we created an issue about coopmat performance, but not about the general performance issues. SYCL is faster, but less flexible than Vulkan.